+++ R/datsearch.r
@@ -72,8 +72,8 @@ In terms of fields/topics, the following terms have been used at least once:
 alternative medicine, attraction, cardiology, climate change, covid-19,
 criminology, dentistry, ecology, education, engineering, epidemiology,
 evolution, genetics, human factors, medicine, memory, obstetrics, oncology,
-persuasion, primary care, psychiatry, psychology, pulmonology, smoking,
-social work, sociology, sports.
+persuasion, primary care, psychiatry, psychology, pulmonology, rheumatology,
+smoking, social work, sociology, sports.
 
 In terms of outcome measures, the following terms have been used at least once:
 
@@ -87,10 +87,10 @@ In terms of models/methods/concepts, the following terms have been used at least
 cluster-robust inference, component network meta-analysis, cumulative
 meta-analysis, diagnostic accuracy studies, dose response models, generalized
 linear models, individual participant data, longitudinal models,
-Mantel-Haenszel method, meta-regression, model checks, multilevel models,
-multivariate models, network meta-analysis, outliers, Peto's method, phylogeny,
-publication bias, reliability generalization, risk-of-bias, single-arm studies,
-spatial correlation, subgroup analysis.")
+Mantel-Haenszel method, meta-regression, missing data, model checks,
+multilevel models, multivariate models, network meta-analysis, outliers,
+Peto's method, phylogeny, prevalence, publication bias, reliability generalization,
+risk-of-bias, single-arm studies, spatial correlation, subgroup analysis.")
          cat("\n\n")
          pattern <- NULL
          next