+++ data-raw/dat.feng2019.r
@@ -4,4 +4,4 @@
 # Medicine, 98.
 
 dat.feng2019 <- read.csv("data-raw/dat.feng2019.csv")
-save(dat.feng2019, file ="data/dat.feng2019.rda")
+save(dat.feng2019, file="data/dat.feng2019.rda")