+++ data-raw/dat.greb2008.r
@@ -5,4 +5,4 @@
 # CD004024.
 
 dat.greb2008 <- read.csv("data-raw/dat.greb2008.csv")
-save(dat.greb2008, file = "data/dat.greb2008.rda")
+save(dat.greb2008, file="data/dat.greb2008.rda")