+++ data-raw/dat.greenland1992.r
@@ -4,4 +4,4 @@
 # American Journal of Epidemiology, 135, 1301-9.
 
 dat.greenland1992 <- read.csv("data-raw/dat.greenland1992.csv")
-save(dat.greenland1992, file ="data/dat.greenland1992.rda")
+save(dat.greenland1992, file="data/dat.greenland1992.rda")