+++ data-raw/dat.higgins2008.r
@@ -4,4 +4,4 @@
 # Clinical Trials, 5 (3): 225-39
 
 dat.higgins2008 <- read.csv("data-raw/dat.higgins2008.csv")
-save(dat.higgins2008, file = "data/dat.higgins2008.rda")
+save(dat.higgins2008, file="data/dat.higgins2008.rda")