+++ data-raw/dat.lloyd2010.r
@@ -4,4 +4,4 @@
 # Rheumatology (Oxford), 49 (12): 2313-21
 
 dat.lloyd2010 <- read.csv("data-raw/dat.lloyd2010.csv")
-save(dat.lloyd2010, file = "data/dat.lloyd2010.rda")
+save(dat.lloyd2010, file="data/dat.lloyd2010.rda")