+++ data-raw/dat.moore1998.r
@@ -1,7 +1,7 @@
-# Lloyd S, Bujkiewicz S, Wailoo AJ, et al. (2010).
-# The effectiveness of anti-TNF-alpha therapies when used sequentially in
-# rheumatoid arthritis patients: A systematic review and meta-analysis.
-# Rheumatology (Oxford), 49 (12): 2313-21
+# Moore, R. A., Tramèr, M. R., Carroll, D., Wiffen, P. J., & McQuay, H. J.
+# (1998). Quantitive systematic review of topically applied non-steroidal
+# anti-inflammatory drugs. British Medical Journal, 316(7128), 333-338.
+# https://doi.org/10.1136/bmj.316.7128.333
 
 dat.moore1998 <- read.csv("data-raw/dat.moore1998.csv")
-save(dat.moore1998, file = "data/dat.moore1998.rda")
+save(dat.moore1998, file="data/dat.moore1998.rda")