+++ data-raw/dat.poole2003.r
@@ -4,4 +4,4 @@
 # Cochrane Database of Systematic Reviews, Issue 1, CD001287
 
 dat.poole2003 <- read.csv("data-raw/dat.poole2003.csv")
-save(dat.poole2003, file ="data/dat.poole2003.rda")
+save(dat.poole2003, file="data/dat.poole2003.rda")