+++ data-raw/dat.steurer2006.r
@@ -1,6 +1,7 @@
-# Quan AP, Kerlikowske K, Gueyffier F, et al., and INDANA Investigators (2000).
-# Pharmacotherapy for hypertension in women of different races.
-# Cochrane Database of Systematic Reviews, 2, CD002146
+# Steurer, M., Pall, G., Richards, S., Schwarzer, G., Bohlius, J., & Greil, R.
+# (2006). Purine antagonists for chronic lymphocytic leukaemia. Cochrane
+# Database of Systematic Reviews, 3, CD004270.
+# https://doi.org/10.1002/14651858.cd004270.pub2
 
 dat.steurer2006 <- read.csv("data-raw/dat.steurer2006.csv")
-save(dat.steurer2006, file ="data/dat.steurer2006.rda")
+save(dat.steurer2006, file="data/dat.steurer2006.rda")